Now it's work, add warp system

+++ src/main/kotlin/moe/muska/ami/spcraft/cartcore/packs/warp/OnCommand.kt
@@ -1,14 +1,19 @@
 package moe.muska.ami.spcraft.cartcore.packs.warp
 
+import moe.muska.ami.spcraft.cartcore.utils.Configuration
 import moe.muska.ami.spcraft.cartcore.utils.Hook
 import moe.muska.ami.spcraft.cartcore.utils.Logger
 import org.bukkit.command.Command
 import org.bukkit.command.CommandSender
+import org.bukkit.configuration.file.FileConfiguration
 import org.bukkit.entity.Player
+import java.math.BigDecimal
+import java.math.RoundingMode
 
 class OnCommand: moe.muska.ami.spcraft.cartcore.utils.Command() {
 
-    val db: DB = DB()
+    private val config: FileConfiguration? = Configuration.PacksConfiguration.WARP
+    private val db: DB = DB()
 
     class OnCommandInject(cmd: String): Command(cmd) {
         override fun execute(sender: CommandSender, p1: String, args: Array<out String>?): Boolean {
@@ -24,17 +29,25 @@ class OnCommand: moe.muska.ami.spcraft.cartcore.utils.Command() {
 
     override fun execute(sender: CommandSender, args: Array<out String>?) {
         //sender.sendMessage("OK")
-        if (args.isNullOrEmpty()) {
-            sender.sendMessage(Hook.translateMiniMessage("<red>命令语法不正确</red>"))
-            return
-        }
-        when (args[0]) {
-            "set" -> setWarp(sender, args.drop(1).toTypedArray())
-            "del" -> deleteWarp(sender, args.drop(1).toTypedArray())
-            "delete" -> deleteWarp(sender, args.drop(1).toTypedArray())
-            "tp" -> teleportWarp(sender, args.drop(1).toTypedArray())
-            "teleport" -> teleportWarp(sender, args.drop(1).toTypedArray())
-            else -> teleportWarp(sender, args)
+        if (config?.getBoolean("enable") == true) {
+            if (sender.hasPermission("cartcore.warp")) {
+                if (args.isNullOrEmpty()) {
+                    sender.sendMessage(Hook.translateMiniMessage("<red>命令语法不正确</red>"))
+                    return
+                }
+                when (args[0]) {
+                    "set" -> setWarp(sender, args.drop(1).toTypedArray())
+                    "del" -> deleteWarp(sender, args.drop(1).toTypedArray())
+                    "delete" -> deleteWarp(sender, args.drop(1).toTypedArray())
+                    "tp" -> teleportWarp(sender, args.drop(1).toTypedArray())
+                    "teleport" -> teleportWarp(sender, args.drop(1).toTypedArray())
+                    else -> teleportWarp(sender, args)
+                }
+            } else {
+                sender.sendMessage(Hook.translateMiniMessage("<red>你没有权限执行此命令</red>"))
+            }
+        } else {
+            sender.sendMessage(Hook.translateMiniMessage("<yellow>功能未启用</yellow>"))
         }
     }
 
@@ -59,9 +72,17 @@ class OnCommand: moe.muska.ami.spcraft.cartcore.utils.Command() {
     private fun teleportWarp(sender: CommandSender, args: Array<out String>?) {
         if (args?.size == 1) {
             val res = db.get(args[0])
-            val player = sender as Player
-            player.teleport(res)
-            sender.sendMessage(Hook.translateMiniMessage("<yellow>已传送至 ${res}</yellow>"))
+            if (res != null) {
+                val player = sender as Player
+                player.teleport(res)
+                val x = BigDecimal(res.x).setScale(2, RoundingMode.HALF_UP)
+                val y = BigDecimal(res.y).setScale(2, RoundingMode.HALF_UP)
+                val z = BigDecimal(res.z).setScale(2, RoundingMode.HALF_UP)
+                val world = res.world.name
+                sender.sendMessage(Hook.translateMiniMessage("<yellow>已传送至 世界: $world, 坐标: $x $y $z</yellow>"))
+            } else {
+                sender.sendMessage(Hook.translateMiniMessage("<red>传送点不存在</red>"))
+            }
         } else {
             sender.sendMessage(Hook.translateMiniMessage("<red>命令语法不正确</red>"))
         }
